Clean code structure
